Allow customer to view promotions in CustomerGUI - update customer perms in db to view promotion - customer can now apply promo codes for discount - customer can now view promotions in gui - add a singleton PromotionManager to be concrete subject of PromotionSubject

+++ src/backend/DatabaseManager.java
@@ -486,7 +486,7 @@ public class DatabaseManager {
 
                 Promotion promotion = new Promotion(
                     rs.getString("promo_code"),
-                    rs.getInt("discount_rate"),
+                    rs.getDouble("discount_rate"),
                     rs.getString("description"),
                     promoStartDate
                 );
